% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{cutOutlierSample}
\alias{cutOutlierSample}
\title{This function plots the samples cluster Tree, helping to find the height value to exclude outiler samples. When height is provided, the function will return a vector of samples to keep.}
\usage{
cutOutlierSample(data, trait, filename, height = FALSE)
}
\description{
This function plots the samples cluster Tree, helping to find the height value to exclude outiler samples. When height is provided, the function will return a vector of samples to keep.
}
