% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{filterBootstrap}
\alias{filterBootstrap}
\title{This function removes low stability genes found in the bootstrap based on the cut threshold}
\usage{
filterBootstrap(allDegrees, bootstrap, stability, cut_threshold)
}
\description{
This function removes low stability genes found in the bootstrap based on the cut threshold
}
