% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{filterTransform}
\alias{filterTransform}
\title{This function filters the expression data (TPM or FPKM), by MAD or DEG, then it applies voom transformation using limma package.}
\usage{
filterTransform(datExpression, datCounts)
}
\arguments{
\item{datExpression}{DEG data}

\item{datCounts}{count data}
}
\description{
This function filters the expression data (TPM or FPKM), by MAD or DEG, then it applies voom transformation using limma package.
}
