% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{heatmapTopConnectivity}
\alias{heatmapTopConnectivity}
\title{This function generates a heatmap of top connected lncRNAs and protein coding genes of a specific submodule.}
\usage{
heatmapTopConnectivity(
  module,
  submodule,
  number_lnc_pc = FALSE,
  TOM,
  datExpr,
  filename,
  traits,
  summList,
  ...
)
}
\description{
This function generates a heatmap of top connected lncRNAs and protein coding genes of a specific submodule.
}
