% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{moduleStability}
\alias{moduleStability}
\title{This function calculates the stability ratio of each gene, and return a graphic of it}
\usage{
moduleStability(bootstrap, filename)
}
\description{
This function calculates the stability ratio of each gene, and return a graphic of it
}
