% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{pickSoftThresholdPlot}
\alias{pickSoftThresholdPlot}
\title{This function generates a image to help the user to select a beta-value that transform the network in a scale-free network}
\usage{
pickSoftThresholdPlot(datExpr, filename, blocksize = 10000)
}
\description{
This function generates a image to help the user to select a beta-value that transform the network in a scale-free network
}
