% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{stackedBarplot}
\alias{stackedBarplot}
\title{This function returns a barplot correlating the eingengene modules with the trait studied}
\usage{
stackedBarplot(summdf, datExpr, traits, filename)
}
\description{
This function returns a barplot correlating the eingengene modules with the trait studied
}
