% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{summarizeData}
\alias{summarizeData}
\title{This function takes several informations and summarizes it in a dataframe}
\usage{
summarizeData(
  allDegrees,
  modules,
  bootstrapStability,
  cutBootstrap,
  ncGeneNames,
  cGeneNames,
  datExpression
)
}
\arguments{
\item{allDegrees}{connectivity}

\item{modules}{gene modules}

\item{bootstrapStability}{module stability genes}

\item{cutBootstrap}{filter module stability genes}

\item{ncGeneNames}{annotation non coding}

\item{cGeneNames}{annotation coding}

\item{datExpression}{DEG data}
}
\description{
This function takes several informations and summarizes it in a dataframe
}
