% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{summarizeSubmodules}
\alias{summarizeSubmodules}
\title{This function adds the reduced enrichment analysis to a summarized dataframe}
\usage{
summarizeSubmodules(summdf, rrvgolist, listgprof)
}
\arguments{
\item{summdf}{summarized table}

\item{rrvgolist}{enrichment summarized table}

\item{listgprof}{enriched table}
}
\description{
This function adds the reduced enrichment analysis to a summarized dataframe
}
