% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{downloadGeneName}
\alias{downloadGeneName}
\title{This function downloads a gtf table of some genome annotation, using a URL link provided by the user}
\usage{
downloadGeneName(link)
}
\arguments{
\item{link}{URL link of an annotation gtf table}
}
\description{
This function downloads a gtf table of some genome annotation, using a URL link provided by the user
}
