% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{downloadNonCoding}
\alias{downloadNonCoding}
\title{This function downloads a gtf table of non coding genes annotation, using a link provided by the user}
\usage{
downloadNonCoding(nclink)
}
\description{
This function downloads a gtf table of non coding genes annotation, using a link provided by the user
}
