% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{enrichList}
\alias{enrichList}
\title{This function does the gprofiler enrichment analysis over each module}
\usage{
enrichList(summdf)
}
\description{
This function does the gprofiler enrichment analysis over each module
}
