% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{filterBootstrap}
\alias{filterBootstrap}
\title{This function removes low stability genes found in the bootstrap based on the cut threshold}
\usage{
filterBootstrap(allDegrees, bootstrap, stability, cut_threshold)
}
\arguments{
\item{allDegrees}{connectivity table}

\item{bootstrap}{bootrastp table}

\item{stability}{stability table}

\item{cut_threshold}{threshold to remove module low stability genes}
}
\description{
This function removes low stability genes found in the bootstrap based on the cut threshold
}
