% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{geneIDtoGeneName}
\alias{geneIDtoGeneName}
\title{This function converts the geneIDs data to gene names}
\usage{
geneIDtoGeneName(allDegreesBootstrap, differentialExpression, ncData)
}
\arguments{
\item{allDegreesBootstrap}{bootstrap selected alldegree table}

\item{differentialExpression}{filtered DE genes}

\item{ncData}{non-coding annotation}
}
\description{
This function converts the geneIDs data to gene names
}
