% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{heatmapTopConnectivity}
\alias{heatmapTopConnectivity}
\title{This function generates a heatmap of top connected lncRNAs and protein coding genes of a specific submodule.}
\usage{
heatmapTopConnectivity(
  module,
  submodule,
  hmDimensions = FALSE,
  datExpr,
  filename = FALSE,
  traits,
  summList,
  removeNonDEG = FALSE,
  outTSV = FALSE,
  ...
)
}
\arguments{
\item{module}{module number}

\item{submodule}{submodule number}

\item{hmDimensions}{dim hm}

\item{datExpr}{filtered count data}

\item{filename}{filename}

\item{traits}{tumor/non-tumor}

\item{summList}{summary list table}

\item{removeNonDEG}{show only DEGs}

\item{outTSV}{generate a tsv table}

\item{...}{others}
}
\description{
This function generates a heatmap of top connected lncRNAs and protein coding genes of a specific submodule.
}
