% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{reduceEnrichment}
\alias{reduceEnrichment}
\title{This function performs the rrvgo enrichment summarization of the module enrichment analysis}
\usage{
reduceEnrichment(listgprof, summdf, datExpr, indicePower)
}
\arguments{
\item{listgprof}{enriched table}

\item{summdf}{summarized table}

\item{datExpr}{count data}

\item{indicePower}{beta-value}
}
\description{
This function performs the rrvgo enrichment summarization of the module enrichment analysis
}
