% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{saveEnrichedGraph}
\alias{saveEnrichedGraph}
\title{This function does the reduction of the module enrichment, and returns a figure}
\usage{
saveEnrichedGraph(rrvgolist, summdf, datExpr, traits, filename)
}
\description{
This function does the reduction of the module enrichment, and returns a figure
}
