% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{summarizeAndEnrichModules}
\alias{summarizeAndEnrichModules}
\title{This function prepares the co-expression network, performs the module enrichment analysis and summarize all the data in a list.}
\usage{
summarizeAndEnrichModules(
  annotationData,
  datExpression,
  ncGeneNames,
  datExpr,
  indicePower,
  traits,
  cutBootstrap = FALSE,
  bootstrapStability = FALSE,
  maxBlockSize = maxBlockSize,
  TOMType = "unsigned",
  minModuleSize = 30,
  reassignThreshold = 0,
  mergeCutHeight = 0.3,
  pamRespectsDendro = FALSE,
  corType = "bicor",
  user_threshold = 0.05,
  ontology = "BP",
  organism = "hsapiens",
  orgdb = "org.Hs.eg.db",
  reducedTermsThreshold = 0.7,
  filename = "5_enrichedgraph.png",
  ...
)
}
\arguments{
\item{annotationData}{gene annotation}

\item{datExpression}{DEG information}

\item{ncGeneNames}{non coding annotation}

\item{datExpr}{dat expression tpm/fpkm}

\item{indicePower}{free scale topology beta-value}

\item{traits}{tumor/non-tumor}

\item{cutBootstrap}{threshhold gene stability in modules}

\item{bootstrapStability}{gene stability table}

\item{maxBlockSize}{wgcna parameter, bigger to faster computers}
}
\description{
This function prepares the co-expression network, performs the module enrichment analysis and summarize all the data in a list.
}
