% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{summarizeAndEnrichModules}
\alias{summarizeAndEnrichModules}
\title{Summarize all data in a list, make the network and do the enrichment analysis}
\usage{
summarizeAndEnrichModules(
  annotationData,
  datExpr,
  indicePower,
  traits,
  cutBootstrap,
  bootstrapStability,
  maxBlockSize = 10000
)
}
\arguments{
\item{annotationData}{annotation data file}

\item{datExpr}{expression file}

\item{indicePower}{beta value}

\item{traits}{trait file}

\item{cutBootstrap}{threshold to remove less stable genes in the network}

\item{bootstrapStability}{file generated by the server script, false if not bootstrapping}

\item{maxBlockSize}{max size processed by wgcna networks}
}
\value{
a list with 1- summarized data table, 2- Enriched data list, 3- Topological overlap table
}
\description{
Summarize all data in a list, make the network and do the enrichment analysis
}
