% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{summarizeData}
\alias{summarizeData}
\title{This function takes several informations and summarizes it in a dataframe}
\usage{
summarizeData(
  allDegrees,
  modules,
  bootstrapStability,
  cutBootstrap,
  ncGeneNames,
  cGeneNames,
  datExpression
)
}
\description{
This function takes several informations and summarizes it in a dataframe
}
